changed M_PI definition to ward off long double problems

+++ src/cv/mgflib/parser.h
@@ -463,7 +463,7 @@ extern int	xf(XF *, int, char **);		/* interpret transform spec. */
 
 #ifndef  PI
 #ifdef	M_PI
-#define	 PI		M_PI
+#define	 PI		((double)M_PI)
 #else
 #define	 PI		3.14159265358979323846
 #endif
